Adding in the rich text element
