+++ src/edit.js
@@ -1,488 +1,115 @@
-import { createElement, Fragment } from "@wordpress/element";
-
-/**
- * External dependencies
- */
-import classnames from 'classnames';
-/**
- * WordPress dependencies
- */
-
-import { useSelect, useDispatch } from '@wordpress/data';
-import { PanelBody, TextControl, TextareaControl, ToolbarButton, ToolbarGroup } from '@wordpress/components';
-import { displayShortcut, isKeyboardEvent } from '@wordpress/keycodes';
+import { SelectControl } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { BlockControls, InnerBlocks, useInnerBlocksProps, InspectorControls, RichText, useBlockProps, store as blockEditorStore, getColorClassName } from '@wordpress/block-editor';
-import { isURL, prependHTTP } from '@wordpress/url';
-import { useState, useEffect, useRef } from '@wordpress/element';
-import { placeCaretAtHorizontalEdge } from '@wordpress/dom';
-import { link as linkIcon, removeSubmenu } from '@wordpress/icons';
-import { useResourcePermissions } from '@wordpress/core-data';
-import { speak } from '@wordpress/a11y';
-import { createBlock } from '@wordpress/blocks';
-import { useMergeRefs, usePrevious } from '@wordpress/compose';
-/**
- * Internal dependencies
- */
-
-import { ItemSubmenuIcon } from '@wordpress/block-library/build-module/navigation-submenu/icons';
-const {
-  name: name
-} = {
-  $schema: "https://schemas.wp.org/trunk/block.json",
-  apiVersion: 2,
-  name: "core/navigation-submenu",
-  title: "Submenu",
-  category: "design",
-  parent: ["core/navigation"],
-  description: "Add a submenu to your navigation.",
-  textdomain: "default",
-  attributes: {
-    label: {
-      type: "string"
-    },
-    type: {
-      type: "string"
-    },
-    description: {
-      type: "string"
-    },
-    rel: {
-      type: "string"
-    },
-    id: {
-      type: "number"
-    },
-    opensInNewTab: {
-      type: "boolean",
-      "default": false
-    },
-    url: {
-      type: "string"
-    },
-    title: {
-      type: "string"
-    },
-    kind: {
-      type: "string"
-    },
-    isTopLevelItem: {
-      type: "boolean",
-	  "default": true,
-    }
-  },
-  usesContext: ["textColor", "customTextColor", "backgroundColor", "customBackgroundColor", "overlayTextColor", "customOverlayTextColor", "overlayBackgroundColor", "customOverlayBackgroundColor", "fontSize", "customFontSize", "showSubmenuIcon", "maxNestingLevel", "openSubmenusOnClick", "style"],
-  supports: {
-    reusable: false,
-    html: false
-  },
-  editorStyle: "wp-block-navigation-submenu-editor",
-  style: "wp-block-navigation-submenu"
-};
 import { LinkUI } from '@wordpress/block-library/build-module/navigation-link/link-ui';
-import { updateAttributes } from '@wordpress/block-library/build-module/navigation-link/update-attributes';
-import { getColors, getNavigationChildBlockProps } from '@wordpress/block-library/build-module/navigation/edit/utils';
-const ALLOWED_BLOCKS = ['core/navigation-link', 'core/navigation-submenu', 'core/page-list'];
-const DEFAULT_BLOCK = {
-  name: 'core/navigation-link'
-};
-/**
- * A React hook to determine if it's dragging within the target element.
- *
- * @typedef {import('@wordpress/element').RefObject} RefObject
- *
- * @param {RefObject<HTMLElement>} elementRef The target elementRef object.
- *
- * @return {boolean} Is dragging within the target element.
- */
+import { useMergeRefs, usePrevious } from '@wordpress/compose';
+import { useState, useEffect, useRef } from '@wordpress/element';
+import { speak } from '@wordpress/a11y';
 
-const useIsDraggingWithin = elementRef => {
-  const [isDraggingWithin, setIsDraggingWithin] = useState(false);
-  useEffect(() => {
-    const {
-      ownerDocument
-    } = elementRef.current;
-
-    function handleDragStart(event) {
-      // Check the first time when the dragging starts.
-      handleDragEnter(event);
-    } // Set to false whenever the user cancel the drag event by either releasing the mouse or press Escape.
-
-
-    function handleDragEnd() {
-      setIsDraggingWithin(false);
-    }
-
-    function handleDragEnter(event) {
-      // Check if the current target is inside the item element.
-      if (elementRef.current.contains(event.target)) {
-        setIsDraggingWithin(true);
-      } else {
-        setIsDraggingWithin(false);
-      }
-    } // Bind these events to the document to catch all drag events.
-    // Ideally, we can also use `event.relatedTarget`, but sadly that
-    // doesn't work in Safari.
-
-
-    ownerDocument.addEventListener('dragstart', handleDragStart);
-    ownerDocument.addEventListener('dragend', handleDragEnd);
-    ownerDocument.addEventListener('dragenter', handleDragEnter);
-    return () => {
-      ownerDocument.removeEventListener('dragstart', handleDragStart);
-      ownerDocument.removeEventListener('dragend', handleDragEnd);
-      ownerDocument.removeEventListener('dragenter', handleDragEnter);
-    };
-  }, []);
-  return isDraggingWithin;
-};
-/**
- * @typedef {'post-type'|'custom'|'taxonomy'|'post-type-archive'} WPNavigationLinkKind
- */
 
 /**
- * Navigation Link Block Attributes
+ * Output our menu selector.
  *
- * @typedef {Object} WPNavigationLinkBlockAttributes
- *
- * @property {string}               [label]         Link text.
- * @property {WPNavigationLinkKind} [kind]          Kind is used to differentiate between term and post ids to check post draft status.
- * @property {string}               [type]          The type such as post, page, tag, category and other custom types.
- * @property {string}               [rel]           The relationship of the linked URL.
- * @property {number}               [id]            A post or term id.
- * @property {boolean}              [opensInNewTab] Sets link target to _blank when true.
- * @property {string}               [url]           Link href.
- * @property {string}               [title]         Link title attribute.
+ * @return {WPElement} Element to render.
  */
+export default function Edit( props ) {
+
+	let {
+		attributes,
+		isSelected,
+		setAttributes,
+		mergeBlocks,
+		onReplace,
+		context,
+		clientId
+	} = props;
+
+	const {
+		label,
+		type,
+		url,
+		description,
+		rel,
+		title
+	} = attributes;
+
+	const {
+		showSubmenuIcon,
+		maxNestingLevel,
+		openSubmenusOnClick
+	} = context;
+
+	const { useSelect } = wp.data;
+	const [isLinkOpen, setIsLinkOpen] = useState(false); // Use internal state instead of a ref to make sure that the component
+	const [popoverAnchor, setPopoverAnchor] = useState(null);
+	const ref = useRef();
+	const itemLabelPlaceholder = __('Add text…');
+
+	// Get the wp_template_parts
+	const templateParts = useSelect( ( select ) => {
+		return select( 'core' ).getEntityRecords( 'postType', 'wp_template_part' )
+	} )
 
 
-export default function NavigationSubmenuEdit(_ref) {
-  let {
-    attributes,
-    isSelected,
-    setAttributes,
-    mergeBlocks,
-    onReplace,
-    context,
-    clientId
-  } = _ref;
+	let html = '';
 
-  console.log(_ref);
-  console.log(context);
+	if ( null === templateParts ) {
+		html = ( <div>Loading</div> );
+	} else {
 
-  const {
-    label,
-    type,
-    url,
-    description,
-    rel,
-    title
-  } = attributes;
-  const {
-    showSubmenuIcon,
-    maxNestingLevel,
-    openSubmenusOnClick
-  } = context;
-  const {
-    __unstableMarkNextChangeAsNotPersistent,
-    replaceBlock
-  } = useDispatch(blockEditorStore);
-  const [isLinkOpen, setIsLinkOpen] = useState(false); // Use internal state instead of a ref to make sure that the component
-  // re-renders when the popover's anchor updates.
+		/*html = createElement(RichText, {
+			ref: ref,
+			identifier: "label",
+			className: "wp-block-navigation-item__label",
+			value: label,
+			onChange: labelValue => setAttributes({
+			  label: labelValue
+			}),
+			"aria-label": __('Navigation link text'),
+			placeholder: itemLabelPlaceholder,
+			withoutInteractiveFormatting: true,
+			allowedFormats: ['core/bold', 'core/italic', 'core/image', 'core/strikethrough'],
+			onClick: () => {
+			  if (! openSubmenusOnClick && !url) {
+				setIsLinkOpen(true);
+			  }
+			}
+		});*/
 
-  const [popoverAnchor, setPopoverAnchor] = useState(null);
-  const listItemRef = useRef(null);
-  const isDraggingWithin = useIsDraggingWithin(listItemRef);
-
-  const itemLabelPlaceholder = __('Add text…');
-
-  const ref = useRef();
-  const pagesPermissions = useResourcePermissions('pages');
-  const postsPermissions = useResourcePermissions('posts');
-  const {
-    isAtMaxNesting,
-    isTopLevelItem,
-    isParentOfSelectedBlock,
-    isImmediateParentOfSelectedBlock,
-    hasChildren,
-    selectedBlockHasChildren,
-    onlyDescendantIsEmptyLink
-  } = useSelect(select => {
-    const {
-      hasSelectedInnerBlock,
-      getSelectedBlockClientId,
-      getBlockParentsByBlockName,
-      getBlock,
-      getBlockCount,
-      getBlockOrder
-    } = select(blockEditorStore);
-
-    let _onlyDescendantIsEmptyLink;
-
-    const selectedBlockId = getSelectedBlockClientId();
-    const selectedBlockChildren = getBlockOrder(selectedBlockId); // Check for a single descendant in the submenu. If that block
-    // is a link block in a "placeholder" state with no label then
-    // we can consider as an "empty" link.
-
-    if ((selectedBlockChildren === null || selectedBlockChildren === void 0 ? void 0 : selectedBlockChildren.length) === 1) {
-      var _singleBlock$attribut;
-
-      const singleBlock = getBlock(selectedBlockChildren[0]);
-      _onlyDescendantIsEmptyLink = (singleBlock === null || singleBlock === void 0 ? void 0 : singleBlock.name) === 'core/navigation-link' && !(singleBlock !== null && singleBlock !== void 0 && (_singleBlock$attribut = singleBlock.attributes) !== null && _singleBlock$attribut !== void 0 && _singleBlock$attribut.label);
-    }
-
-    return {
-      isAtMaxNesting: getBlockParentsByBlockName(clientId, name).length >= maxNestingLevel,
-      isTopLevelItem: getBlockParentsByBlockName(clientId, name).length === 0,
-      isParentOfSelectedBlock: hasSelectedInnerBlock(clientId, true),
-      isImmediateParentOfSelectedBlock: hasSelectedInnerBlock(clientId, false),
-      hasChildren: !!getBlockCount(clientId),
-      selectedBlockHasChildren: !!(selectedBlockChildren !== null && selectedBlockChildren !== void 0 && selectedBlockChildren.length),
-      onlyDescendantIsEmptyLink: _onlyDescendantIsEmptyLink
-    };
-  }, [clientId]);
-  const prevHasChildren = usePrevious(hasChildren); // Show the LinkControl on mount if the URL is empty
-  // ( When adding a new menu item)
-  // This can't be done in the useState call because it conflicts
-  // with the autofocus behavior of the BlockListBlock component.
-
-  useEffect(() => {
-    if (!openSubmenusOnClick && !url) {
-      setIsLinkOpen(true);
-    }
-  }, []); // Store the colors from context as attributes for rendering.
-
-  useEffect(() => {
-    // This side-effect should not create an undo level as those should
-    // only be created via user interactions. Mark this change as
-    // not persistent to avoid undo level creation.
-    // See https://github.com/WordPress/gutenberg/issues/34564.
-    __unstableMarkNextChangeAsNotPersistent();
-
-    setAttributes({
-      isTopLevelItem
-    });
-  }, [isTopLevelItem]);
-  /**
-   * The hook shouldn't be necessary but due to a focus loss happening
-   * when selecting a suggestion in the link popover, we force close on block unselection.
-   */
-
-  useEffect(() => {
-    if (!isSelected) {
-      setIsLinkOpen(false);
-    }
-  }, [isSelected]); // If the LinkControl popover is open and the URL has changed, close the LinkControl and focus the label text.
-
-  useEffect(() => {
-    if (isLinkOpen && url) {
-      // Does this look like a URL and have something TLD-ish?
-      if (isURL(prependHTTP(label)) && /^.+\.[a-z]+/.test(label)) {
-        // Focus and select the label text.
-        selectLabelText();
-      } else {
-        // Focus it (but do not select).
-        placeCaretAtHorizontalEdge(ref.current, true);
-      }
-    }
-  }, [url]);
-  /**
-   * Focus the Link label text and select it.
-   */
-
-  function selectLabelText() {
-    ref.current.focus();
-    const {
-      ownerDocument
-    } = ref.current;
-    const {
-      defaultView
-    } = ownerDocument;
-    const selection = defaultView.getSelection();
-    const range = ownerDocument.createRange(); // Get the range of the current ref contents so we can add this range to the selection.
-
-    range.selectNodeContents(ref.current);
-    selection.removeAllRanges();
-    selection.addRange(range);
-  }
-
-  let userCanCreate = false;
-
-  if (!type || type === 'page') {
-    userCanCreate = pagesPermissions.canCreate;
-  } else if (type === 'post') {
-    userCanCreate = postsPermissions.canCreate;
-  }
-
-  const {
-    textColor,
-    customTextColor,
-    backgroundColor,
-    customBackgroundColor
-  } = getColors(context, !isTopLevelItem);
-
-  function onKeyDown(event) {
-    if (isKeyboardEvent.primary(event, 'k')) {
-      setIsLinkOpen(true);
-    }
-  }
-
-  const blockProps = useBlockProps({
-    ref: useMergeRefs([setPopoverAnchor, listItemRef]),
-    className: classnames('wp-block-navigation-item', {
-      'is-editing': isSelected || isParentOfSelectedBlock,
-      'is-dragging-within': isDraggingWithin,
-      'has-link': !!url,
-      'has-child': hasChildren,
-      'has-text-color': !!textColor || !!customTextColor,
-      [getColorClassName('color', textColor)]: !!textColor,
-      'has-background': !!backgroundColor || customBackgroundColor,
-      [getColorClassName('background-color', backgroundColor)]: !!backgroundColor,
-      'open-on-click': openSubmenusOnClick
-    }),
-    style: {
-      color: !textColor && customTextColor,
-      backgroundColor: !backgroundColor && customBackgroundColor
-    },
-    onKeyDown
-  }); // Always use overlay colors for submenus.
-
-  const innerBlocksColors = getColors(context, true);
-  const allowedBlocks = isAtMaxNesting ? ALLOWED_BLOCKS.filter(blockName => blockName !== 'core/navigation-submenu') : ALLOWED_BLOCKS;
-  const navigationChildBlockProps = getNavigationChildBlockProps(innerBlocksColors);
-  const innerBlocksProps = useInnerBlocksProps(navigationChildBlockProps, {
-    allowedBlocks,
-    __experimentalDefaultBlock: DEFAULT_BLOCK,
-    __experimentalDirectInsert: true,
-    // Ensure block toolbar is not too far removed from item
-    // being edited.
-    // see: https://github.com/WordPress/gutenberg/pull/34615.
-    __experimentalCaptureToolbars: true,
-    renderAppender: isSelected || isImmediateParentOfSelectedBlock && !selectedBlockHasChildren || // Show the appender while dragging to allow inserting element between item and the appender.
-    hasChildren ? InnerBlocks.ButtonBlockAppender : false
-  });
-  const ParentElement = openSubmenusOnClick ? 'button' : 'a';
-
-  function transformToLink() {
-    const newLinkBlock = createBlock('core/navigation-link', attributes);
-    replaceBlock(clientId, newLinkBlock);
-  }
-
-  useEffect(() => {
-    // If block becomes empty, transform to Navigation Link.
-    if (!hasChildren && prevHasChildren) {
-      // This side-effect should not create an undo level as those should
-      // only be created via user interactions.
-      __unstableMarkNextChangeAsNotPersistent();
-
-      transformToLink();
-    }
-  }, [hasChildren, prevHasChildren]);
-  const canConvertToLink = !selectedBlockHasChildren || onlyDescendantIsEmptyLink;
-  return createElement(Fragment, null, createElement(BlockControls, null, createElement(ToolbarGroup, null, !openSubmenusOnClick && createElement(ToolbarButton, {
-    name: "link",
-    icon: linkIcon,
-    title: __('Link'),
-    shortcut: displayShortcut.primary('k'),
-    onClick: () => setIsLinkOpen(true)
-  }), createElement(ToolbarButton, {
-    name: "revert",
-    icon: removeSubmenu,
-    title: __('Convert to Link'),
-    onClick: transformToLink,
-    className: "wp-block-navigation__submenu__revert",
-    isDisabled: !canConvertToLink
-  }))), createElement(InspectorControls, null, createElement(PanelBody, {
-    title: __('Link settings')
-  }, createElement(TextControl, {
-    __nextHasNoMarginBottom: true,
-    value: label || '',
-    onChange: labelValue => {
-      setAttributes({
-        label: labelValue
-      });
-    },
-    label: __('Label'),
-    autoComplete: "off"
-  }), createElement(TextControl, {
-    __nextHasNoMarginBottom: true,
-    value: url || '',
-    onChange: urlValue => {
-      setAttributes({
-        url: urlValue
-      });
-    },
-    label: __('URL'),
-    autoComplete: "off"
-  }), createElement(TextareaControl, {
-    __nextHasNoMarginBottom: true,
-    value: description || '',
-    onChange: descriptionValue => {
-      setAttributes({
-        description: descriptionValue
-      });
-    },
-    label: __('Description'),
-    help: __('The description will be displayed in the menu if the current theme supports it.')
-  }), createElement(TextControl, {
-    __nextHasNoMarginBottom: true,
-    value: title || '',
-    onChange: titleValue => {
-      setAttributes({
-        title: titleValue
-      });
-    },
-    label: __('Link title'),
-    autoComplete: "off"
-  }), createElement(TextControl, {
-    __nextHasNoMarginBottom: true,
-    value: rel || '',
-    onChange: relValue => {
-      setAttributes({
-        rel: relValue
-      });
-    },
-    label: __('Link rel'),
-    autoComplete: "off"
-  }))), createElement("div", blockProps, createElement(ParentElement, {
-    className: "wp-block-navigation-item__content"
-  }, createElement(RichText, {
-    ref: ref,
-    identifier: "label",
-    className: "wp-block-navigation-item__label",
-    value: label,
-    onChange: labelValue => setAttributes({
-      label: labelValue
-    }),
-    onMerge: mergeBlocks,
-    onReplace: onReplace,
-    "aria-label": __('Navigation link text'),
-    placeholder: itemLabelPlaceholder,
-    withoutInteractiveFormatting: true,
-    allowedFormats: ['core/bold', 'core/italic', 'core/image', 'core/strikethrough'],
-    onClick: () => {
-      if (!openSubmenusOnClick && !url) {
-        setIsLinkOpen(true);
-      }
-    }
-  }), !openSubmenusOnClick && isLinkOpen && createElement(LinkUI, {
-    className: "wp-block-navigation-link__inline-link-input",
-    clientId: clientId,
-    link: attributes,
-    onClose: () => setIsLinkOpen(false),
-    anchor: popoverAnchor,
-    hasCreateSuggestion: userCanCreate,
-    onRemove: () => {
-      setAttributes({
-        url: ''
-      });
-      speak(__('Link removed.'), 'assertive');
-    },
-    onChange: updatedValue => {
-      updateAttributes(updatedValue, setAttributes, attributes);
-    }
-  })), (showSubmenuIcon || openSubmenusOnClick) && createElement("span", {
-    className: "wp-block-navigation__submenu-icon"
-  }, createElement(ItemSubmenuIcon, null)), createElement("div", innerBlocksProps)));
+		/*let options = [];
+		options.push( { value: '', label: 'None' } );
+		templateParts.forEach( ( part ) => {
+			if ( -1 < part.slug.search('mega-menu')  ) {
+				options.push( { value : part.slug, label : part.title.rendered } )
+			}
+		});
+		html = (
+			<SelectControl
+				label={ 'Select a mega menu:' }
+				value={ props.attributes.menu }
+				onChange={ ( menu ) => {
+					props.setAttributes( { menu } );
+				} }
+				options={ options }
+			/>
+		);*/
+	}
+	return createElement(RichText, {
+		ref: ref,
+		identifier: "label",
+		className: "wp-block-navigation-item__label",
+		value: label,
+		onChange: labelValue => setAttributes({
+		  label: labelValue
+		}),
+		"aria-label": __('Navigation link text'),
+		placeholder: itemLabelPlaceholder,
+		withoutInteractiveFormatting: true,
+		allowedFormats: ['core/bold', 'core/italic', 'core/image', 'core/strikethrough'],
+		onClick: () => {
+		  if (! openSubmenusOnClick && !url) {
+			setIsLinkOpen(true);
+		  }
+		}
+	});
 }
-//# sourceMappingURL=edit.js.map
